add ClientService interface and the impl
